+++ CHANGELOG.md
@@ -4,7 +4,11 @@
   - Added Localization where possible for some names
   - `Hide Token Effects` updated to cover edge case of effects with a duration in `rounds` not passing through
 
-## 1.8.8
+## 1.9.0
+
+- **New**
+  - `Item Mystification Helper` - Adds an option to drop an item on the identification tab to transform the false identification into that item
+  - Additionally it also will remove the cursed trait on items that are mystified and re-add it when the item is identified
 - **Updated**
   - Updated Chinese translation (🌐 @AlphaStarguide)